refactor: colocate reward manager tests
