+++ components/rewards/reward-manager/redemption-list.history.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import { RedemptionList } from "../redemption-list";
+import { RedemptionList } from "./redemption-list";
 import {
   createMockRedemption,
   createRedemptionHandlers,
