+++ components/rewards/reward-manager/reward-form.interactions.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
-import { RewardForm } from "../reward-form";
+import { RewardForm } from "./reward-form";
 import {
   defaultFormData,
   filledFormData,
